+++ src/main/java/com/proyectocurso/seguimiento/repository/DatosSuscriptorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.r2dbc.repository.R2dbcRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DatosSuscriptorRepository extends R2dbcRepository<DatosSuscriptor, Long> {
+public interface DatosSuscriptorRepository extends R2dbcRepository<DatosSuscriptor, Integer> {
 }
