+++ src/main/java/com/proyectocurso/seguimiento/repository/SuscripcionRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.r2dbc.repository.R2dbcRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SuscripcionRepositorio extends R2dbcRepository<Suscripcion, Long> {
+public interface SuscripcionRepositorio extends R2dbcRepository<Suscripcion, Integer> {
 }
